Created program of tower of Hanoi using recursion.
Tower of Hanoi is a very old problem. People had been solving it for many years. In this problem, you have to shift disk from one peg to another peg using the third peg. Let me explain it more.

The set up of the problem is like in the above picture. There will be three pegs and every peg has its own name as the first peg is named “source” and the second peg is named “destination” and the third peg named as “via”. Now, these named suggest there work.  You have to shift all the disks from source to destination using the via a peg. Now you might be thinking this is the easiest problem you have heard off, you can just pick all the disks and put it in the destination peg without using any third peg. But it’s not that simple because there is catch here, you can move one disk at a time and no larger disk can be above the smaller disk while performing shifting. Now the problem becomes a little tricky.